+++ tsd/tsd-aggregator/src/main/java/com/arpnetworking/tsdaggregator/model/querylog/Version2d.java
@@ -166,7 +166,7 @@ public final class Version2d {
 
         private Sample(final Builder builder) {
             _unit = Optional.fromNullable(builder._unit);
-            _value = builder._value.doubleValue();
+            _value = builder._value;
         }
 
         private final Optional<Unit> _unit;
